+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- A new `CloudEvent` optional `pydantic` model class is available in the
+  `cloudevents.pydantic.event` module. The new model enables the integration of
+   CloudEvents in your existing pydantic models or integration with pydantic
+    dependent systems such as FastAPI. ([#182])
+
+### Changed
+- Deprecated `cloudevents.http.event_type` module,
+    moved under `cloudevents.sdk.converters`.
+- Deprecated `cloudevents.http.json_methods` module,
+    moved under `cloudevents.http.conversion`.
+- Deprecated `cloudevents.http.http_methods` module,
+    moved under `cloudevents.http.conversion`.
+- Deprecated `cloudevents.http.util` module.
+
+
 
 ## [1.5.0] — 2022-08-06
 ### Added
@@ -180,5 +196,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#172]: https://github.com/cloudevents/sdk-python/pull/172
 [#173]: https://github.com/cloudevents/sdk-python/pull/173
 [#180]: https://github.com/cloudevents/sdk-python/pull/180
+[#182]: https://github.com/cloudevents/sdk-python/pull/182
 [#184]: https://github.com/cloudevents/sdk-python/pull/184
 [#186]: https://github.com/cloudevents/sdk-python/pull/186